refactor: improve readibility of setup function
