+++ src/setup/utils.ts
@@ -1,0 +1,3 @@
+export function createRequiredError(field: string) {
+  throw new Error(`[nuxt-auth] config option '${field}' is required`)
+}
